doc: jsdoc work in progress
