+++ app/_components/ui/CallToAction/CallToAction.tsx
@@ -1,9 +1,34 @@
 'use client';
-import React from 'react';
+import dynamic from 'next/dynamic';
+import React, { useEffect, useState } from 'react';
 import { Button } from '@/lib/components/ui/button';
 import CardProps from '@/types/CardProps';
 import { dynamicDownload } from '@/utils/dynamicDownload';
 import { IconLoader } from '@/hooks/IconLoader';
+
+/**
+ * @file CallToAction.tsx
+ * @description This component renders call-to-action buttons with dynamic icons.
+ */
+
+/**
+ * CallToAction component props.
+ * @typedef {Object} CardProps
+ * @property {string} cta1 - The text for the first call-to-action button.
+ * @property {string} icon1 - The icon for the first call-to-action button.
+ * @property {string} href1 - The href for the first call-to-action button.
+ * @property {boolean} downloadActive1 - Whether the first call-to-action button should trigger a download.
+ * @property {string} cta2 - The text for the second call-to-action button.
+ * @property {string} icon2 - The icon for the second call-to-action button.
+ * @property {string} href2 - The href for the second call-to-action button.
+ * @property {boolean} downloadActive2 - Whether the second call-to-action button should trigger a download.
+ */
+
+/**
+ * CallToAction component.
+ * @param {CardProps} props - The props for the component.
+ * @returns {JSX.Element} The rendered component.
+ */
 
 export const CallToAction: React.FC<CardProps> = ({
   cta1,
