+++ app/_components/ui/Card/CardProjects.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import React from 'react';
+import { ExternalLink, FolderGit2, SquarePlay } from 'lucide-react';
 import {
   CardTitle,
   CardDescription,
@@ -11,18 +12,29 @@ import {
 import { Badge } from '@/lib/components/ui/badge';
 import { ScrollArea } from '@/lib/components/ui/scroll-area';
 import { cn } from '@/lib/utils';
-import { Projects } from '@/types/Projects';
-import { ExternalLink, FolderGit2, SquarePlay } from 'lucide-react';
-import { ImageSection } from './CardImage';
+
+import { CardImage } from './CardImage';
 import { dynamicMarginBottom } from '@/utils/dynamicMarginBottom';
 import { baseUrl } from '@/utils/constants/baseUrl';
 
-interface CardProjectsProps {
-  projects: Projects[];
-  className?: string;
-}
+import { Projects } from '@/types/Projects';
 
-export const CardProjects: React.FC<CardProjectsProps> = ({
+/**
+ * @file CardProjects.tsx
+ * @description This file exports a component that renders a list of project cards.
+ */
+
+/**
+ * CardProjects component.
+ * @param {Object} props - The props for the component.
+ * @param {Projects[]} props.projects - An array of project objects to be displayed.
+ * @param {string} [props.className] - Additional class names to apply to the component.
+ * @returns {JSX.Element} The rendered CardProjects component.
+ * @example
+ * <CardProjects projects={projects} className="custom-class" />
+ */
+
+export const CardProjects: React.FC<ProjectsProps> = ({
   projects,
   className,
 }) => {
@@ -44,7 +56,7 @@ export const CardProjects: React.FC<CardProjectsProps> = ({
                 rel='noopener noreferrer'
                 className='flex items-center'
               >
-                <ImageSection
+                <CardImage
                   imageSrc={project.imageSrc}
                   imageAlt={project.imageAlt || ''}
                   width={590}
