+++ app/_components/ui/Section/Section.tsx
@@ -1,6 +1,22 @@
 import { PropsWithChildren } from 'react';
-
 import { cn } from '@/lib/utils';
+
+/**
+ * @file Section.tsx
+ * @description This file exports a component that renders a section with optional additional class names.
+ */
+
+/**
+ * Section component.
+ * @param {PropsWithChildren<{ className?: string }>} props - The props for the component.
+ * @param {React.ReactNode} props.children - The children elements to be rendered inside the section.
+ * @param {string} [props.className] - Additional class names to apply to the section.
+ * @returns {JSX.Element} The rendered section component.
+ * @example
+ * <Section className="custom-class">
+ *   <p>Content goes here</p>
+ * </Section>
+ */
 
 export const Section = ({
   children,
