+++ app/hooks/IconLoader.tsx
@@ -1,5 +1,18 @@
 import React, { useEffect, useState } from 'react';
 
+/**
+ * @file IconLoader.tsx
+ * @description This file exports a component that dynamically loads and renders an icon from the lucide-react library.
+ */
+
+/**
+ * IconLoader component.
+ * @param {string} icon - The name of the icon to load.
+ * @returns {JSX.Element | null} The rendered icon component or null if the icon is not yet loaded.
+ * @example
+ * <IconLoader icon="Home" />
+ */
+
 export const IconLoader = (icon: string) => {
   const [Icon, setIcon] = useState<React.ComponentType | null>(null);
