+++ app/hooks/LocationEventsMap.tsx
@@ -2,6 +2,17 @@ import React, { useEffect } from 'react';
 import { useMap } from 'react-leaflet';
 import { LocationEventsMapProps } from '@/types/LocationEventsProps';
 
+/**
+ * @file LocationEvents.tsx
+ * @description This component handles map events such as flying to a specific position.
+ */
+
+/**
+ * LocationEvents component.
+ * @param {LocationEventsProps} props - The props for the component.
+ * @returns {null} The component does not render anything.
+ */
+
 const LocationEventsMap: React.FC<LocationEventsMapProps> = ({ position }) => {
   const map = useMap();
