--- app/types/Projects.tsx
@@ -1,10 +1,0 @@
-export interface Projects {
-  title: string;
-  organization: string;
-  imageSrc: string;
-  imageAlt: string;
-  demo?: string;
-  website: string;
-  github: string;
-  tags: string[];
-}
