--- app/types/Technologies.tsx
@@ -1,4 +1,0 @@
-export interface Technologies {
-  name: string;
-  slug: string;
-}
