+++ app/utils/constants/baseUrl.tsx
@@ -1,1 +1,13 @@
+/**
+ * @file baseUrl.tsx
+ * @description This file exports a constant that represents the base URL for external links.
+ */
+
+/**
+ * The base URL for external links.
+ * @type {string}
+ * @example
+ * const fullUrl = `${baseUrl}example.com`;
+ */
+
 export const baseUrl = 'https://';
