+++ app/utils/constants/positions.tsx
@@ -1,0 +1,31 @@
+/**
+ * @file positions.tsx
+ * @description This file exports constants that represent geographical positions.
+ */
+
+/**
+ * The geographical position of France.
+ * @type {[number, number]}
+ * @example
+ * const position = francePosition; // [46.6034, 3.1236]
+ */
+
+export const francePosition: [number, number] = [46.6034, 3.1236];
+
+/**
+ * My geographical position.
+ * @type {[number, number]}
+ * @example
+ * const position = myPosition; // [50.381645, 3.053234]
+ */
+
+export const myPosition: [number, number] = [50.381645, 3.053234];
+
+/**
+ * The radius for a geographical area, in meters.
+ * @type {number}
+ * @example
+ * const areaRadius = radius; // 30000
+ */
+
+export const radius = 60000 / 2;
