+++ app/utils/dynamicYear.tsx
@@ -1,3 +1,12 @@
+/**
+ * @file dynamicYear.tsx
+ * @description This file exports the current year dynamically.
+ */
+
+/**
+ * Get the current year.
+ * @returns {number} The current year.
+ */
 const year = new Date().getFullYear();
 
 export default year;
